plots from asset option: toggle 'use_geo_id_input' in paramters/config_runtime.py; fixed integer limitation for all countries over 127

+++ datasets/reproject_to_template.py
@@ -8,6 +8,6 @@ def reproject_to_template(rasterised_vector,template_image):
     output_image = rasterised_vector.reproject(
       crs= crs_template,
       scale= get_scale_from_image(template_image),
-    ).int8()
+    ).int32()
     
     return output_image
